move flexible block to ui folder
