+++ src/components/repromodel-structure/repromodel-structure.jsx
@@ -1,5 +1,5 @@
 import choices from "../../../repromodel_core/choices.json"
-import FlexibleBlock from "../flexible-block/flexible-block"
+import FlexibleBlock from "../ui/flexible-block/flexible-block"
 import React from "react"
 
 import { Container, Stack, Typography } from "@mui/material"
