+++ src/components/ui/flexible-block/flexible-block.jsx
@@ -2,7 +2,7 @@ import "./flexible-block.css"
 
 import React from "react"
 
-import { capitalizeFirstLetter } from "../../utils/string-helpers"
+import { capitalizeFirstLetter } from "../../../utils/string-helpers"
 import { Paper, Typography } from "@mui/material"
 
 function ActiveBlock({ id, name, status }) {
